+++ src/fidokey/pin/client_pin_command.rs
