+++ src/fidokey/pin/mod.rs
@@ -56,3 +56,30 @@ impl FidoKeyHid {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::HidParam;
+    use crate::fidokey::FidoKeyHid;
+    use crate::ctaphid;
+    use crate::Cfg;
+    use super::SubCommand as PinCmd;
+
+    #[test]
+    fn test_client_pin_get_keyagreement() {
+        let hid_params = HidParam::get();
+        let device = FidoKeyHid::new(&hid_params, &Cfg::init()).unwrap();
+        let cid = ctaphid::ctaphid_init(&device).unwrap();
+
+        let send_payload = create_payload(PinCmd::GetKeyAgreement).unwrap();
+        let response_cbor = ctaphid::ctaphid_cbor(&device, &cid, &send_payload).unwrap();
+
+        let key_agreement =
+            client_pin_response::parse_cbor_client_pin_get_keyagreement(&response_cbor).unwrap();
+        println!("authenticatorClientPIN (0x06) - getKeyAgreement");
+        println!("{}", key_agreement);
+
+        assert!(true);
+    }
+}
